Extract ColumnWrapper into its own file
